Abstracted out decoding on sequences
